LENGTH var cast to int

+++ userSettings.py
@@ -1,2 +1,2 @@
-LENGTH = 60*0.5
+LENGTH = int(60*0.5)
 		#60*#of minutes
